requirements.txt
ansicolors
gql
graphql-core
requests
rosdep
python-decouple
requests_toolbelt
soupsieve
bs4
zipp
pyjwt
psycopg2-binary
urllib3
InquirerPy
GitPython
jsonschema
cryptography
numpy
psutil
cmakeast
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
pydantic==1.10.12
python-on-whales
importlib-resources
citros-data-analysis
path
watchdog
rich
config
python-decouple
fastapi[all]
uvicorn
fastapi-utils
python-dotenv
rich_argparse
glob2